CENTRAL/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
certifi==2024.8.30
oracledb==2.4.1
dnspython==2.6.1
